APF-1316
I have completed the update of bootstraps for struts2.
I've done it like this as it includes the bootstrap gif images for
glyphs, as they must be in a certain directory structure so they are
picked up.

It solves as well other issues, at least for Struts2:

http://issues.appfuse.org/browse/APF-1287
http://issues.appfuse.org/browse/APF-1276

and
http://issues.appfuse.org/browse/APF-1275 would be obsolete, if all
frameworks are moved to bootstrap, that is.

As an extra, the focus on the first form field is back, with a jquery
function.