Update LabelTag.java
Matt,

I noticed in our environment LabelTag was taking up a lot of memory, I added the doEndTag method to clear the requestContext which was what was taking up the memory in each user thread on the app server. For some reason just adding requestContext = null to the release method didn't work, I actually had to add the doEndTag method, after doing this the mem usaged dropped, I didn't notice any bugs as a result of this change.